backend/requirements/common.txt
django~=3.1.5
djangorestframework~=3.12.2
djangorestframework-jwt
django-cors-headers
pillow
requests